asynq::components::aggregator

Trait GroupAggregator

pub trait GroupAggregator: Send + Sync {
    // Required method
    fn aggregate(&self, group: &str, tasks: Vec<Task>) -> Result<Task>;
}

GroupAggregator trait - 聚合一组任务为一个任务 GroupAggregator trait - aggregates a group of tasks into one task

对应 Go asynq 的 GroupAggregator 接口 Corresponds to Go asynq’s GroupAggregator interface

type GroupAggregator interface {
    Aggregate(group string, tasks []*Task) *Task
}

Required Methods

fn aggregate(&self, group: &str, tasks: Vec<Task>) -> Result<Task>

聚合给定组中的任务，返回聚合后的新任务 Aggregates the given tasks in a group and returns a new aggregated task

Arguments

group - 组名 / Group name

tasks - 要聚合的任务列表 / List of tasks to aggregate

Returns

聚合后的新任务。队列选项会被忽略，聚合任务总是入队到组所属的队列 The aggregated new task. Queue option will be ignored, aggregated task always enqueues to the group’s queue