asynq::components

Module janitor

Janitor 模块 Janitor module

对应 Go 版本的 janitor.go 职责： Responsibilities corresponding to the Go version’s janitor.go: 定期清理过期的已完成任务、死亡服务器状态等 Periodically clean up expired completed tasks, dead server states, etc.

参考 Go asynq/janitor.go Reference: Go asynq/janitor.go

Structs

Janitor

Janitor - 负责定期清理过期任务和死亡服务器 Janitor - responsible for periodically cleaning up expired tasks and dead servers

JanitorConfig

Janitor 配置 Janitor configuration